requirements.txt
Flask==3.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.25
pyodbc==5.0.1
Pillow==10.2.0
ua-parser==0.18.0
user-agents==2.2.0
python-dateutil==2.8.2